Cleanup

+++ client/src/pages/Post/Post.tsx
@@ -11,7 +11,7 @@ import { useSelector } from "react-redux";
 import { AppDispatch, RootState } from "../../store/store";
 import { useDispatch } from "react-redux";
 import { deletePost } from "../../store/slices/postSlice";
-import Dialog, { TDialog } from "../../components/Dialog";
+import Dialog, { TDialogProps } from "../../components/Dialog";
 
 const Post = () => {
   const dispatch = useDispatch<AppDispatch>();
@@ -21,21 +21,34 @@ const Post = () => {
   const isAuthenticated = useSelector((state: RootState) => state.auth.isAuth);
   const user = useSelector((state: RootState) => state.auth.user);
   const [openDialog, setOpenDialog] = useState(false);
-  const [dialogContent, setDialogContent] = useState<TDialog>({
+  const [dialogContent, setDialogContent] = useState<TDialogProps>({
     title: "",
     description: "",
+    onClose: undefined,
+    onConfirm: undefined,
   });
 
   const handleEdit = () => {
     navigate("edit");
   };
+
   const handleUnpublish = () => {};
+
+  const confirmUnpublish = () => {
+    setDialogContent({
+      title: "Unpublish",
+      description: "Are you sure you want to unpublish this post?",
+      onClose: handleClose,
+      onConfirm: handleUnpublish,
+    });
+    setOpenDialog(true);
+  };
 
   const handleClose = () => {
     setOpenDialog(false);
   };
   const handleDelete = () => {
-    if (post && (user?.id === post?.authorid || user?.admin)) {
+    if (post && user && (user.id === post.authorid || user.admin)) {
       dispatch(deletePost(post.postid));
     }
     setOpenDialog(false);
@@ -45,20 +58,14 @@ const Post = () => {
     setDialogContent({
       title: "Delete post",
       description: "Are you sure you want to delete this post?",
+      onClose: handleClose,
+      onConfirm: handleDelete,
     });
     setOpenDialog(true);
   };
 
   return (
     <article className="post-container">
-      {openDialog && (
-        <Dialog
-          title={dialogContent.title}
-          description={dialogContent.description}
-          onClose={handleClose}
-          onConfirm={handleDelete}
-        />
-      )}
       <div className="post">
         <div className="post-wrapper">
           <div className="post-content">
@@ -69,17 +76,30 @@ const Post = () => {
             user &&
             isAuthenticated &&
             (user.admin || user.id === post.authorid) && (
-              <PostSidebar>
-                <button className="btn edit-btn" onClick={handleEdit}>
-                  Edit
-                </button>
-                <button className="btn unpublish-btn" onClick={handleUnpublish}>
-                  Unpublish
-                </button>
-                <button className="btn delete-btn" onClick={confirmDelete}>
-                  Delete
-                </button>
-              </PostSidebar>
+              <>
+                {openDialog && (
+                  <Dialog
+                    title={dialogContent.title}
+                    description={dialogContent.description}
+                    onClose={dialogContent.onClose}
+                    onConfirm={dialogContent.onConfirm}
+                  />
+                )}
+                <PostSidebar>
+                  <button className="btn edit-btn" onClick={handleEdit}>
+                    Edit
+                  </button>
+                  <button
+                    className="btn unpublish-btn"
+                    onClick={confirmUnpublish}
+                  >
+                    Unpublish
+                  </button>
+                  <button className="btn delete-btn" onClick={confirmDelete}>
+                    Delete
+                  </button>
+                </PostSidebar>
+              </>
             )}
         </div>
       </div>
